Refactor: Move loader function to loader module
Move showLoader and hideLoader function from utility to loader module.
