+++ src/components/Loader/Loader.ts
@@ -1,7 +1,14 @@
 import './Loader.scss'
-import { hideLoader } from '../../ts/_utils'
+
+function showLoader() {
+  loader.classList.remove('loader-container--hide');
+}
+
+function hideLoader() {
+  loader.classList.add('loader-container--hide');
+}
 
 const loader = document.getElementById('loader-container') as HTMLDivElement
 window.onload = () => hideLoader()
 
-export { loader }
+export { loader, showLoader, hideLoader }
