+++ src/ts/components/_airtable.ts
@@ -2,7 +2,8 @@ import { Base } from "airtable";
 import { getDatabaseRecords } from "../../components/_utils";
 import { MovieDetails } from "../../components/Movie/Movie.interface";
 import { renderMovies } from "../../components/Movie/Movie";
-import { capitalize, hideLoader, getAirTableBase } from "../_utils";
+import { capitalize, getAirTableBase } from "../_utils";
+import { hideLoader } from "../../components/Loader/Loader";
 import { modal, showModal, autoCloseModal } from "../../components/Modal/Modal";
 import {
   movieDBSaveMarkup,
